fixed argument handling
the `position_check` function was updated to return three arguments but
two of the calls in the code were still only looking for two items in the
return tuple. Those both have been updated. (The update involved simply
dropping the return as it is not needed in either case; a follow-up issue
might be to further refine so that only needed return values are provided.)